% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NTE.depStartEnd.R
\name{NTE.depStartEnd}
\alias{NTE.depStartEnd}
\title{guess the deposition start and end time}
\usage{
NTE.depStartEnd(d, buffer = 0)
}
\arguments{
\item{d}{data frame with NTE data}

\item{buffer}{how much data to show before / after plot, 1=deposition time}
}
\value{
data frame
}
\description{
guess the deposition start and end time
}
\examples{
d = NTE.load('20191121.txt')
g1 = NTE.plotDeposition(d)
}
