% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NTE.getParams.R
\name{NTE.getParams}
\alias{NTE.getParams}
\title{return average parameters from deposition}
\usage{
NTE.getParams(d)
}
\arguments{
\item{d}{data frame with NTE data}
}
\value{
data frame
}
\description{
return average parameters from deposition
}
\examples{
d = NTE.load('20191121.txt')
NTE.getParams(d)
}
