% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NTE.load.R
\name{NTE.load}
\alias{NTE.load}
\title{loads the data from an NTE-3000 generated data file
into a data frame}
\usage{
NTE.load(filename)
}
\arguments{
\item{filename}{filename including path}
}
\value{
data.frame
}
\description{
loads the data from an NTE-3000 generated data file
into a data frame
}
\examples{
d = NTE.load('20191121.txt')
}
