% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NTE.plotDeposition.R
\name{NTE.plotDeposition}
\alias{NTE.plotDeposition}
\title{graph of the deposition, colors the actual deposition portion}
\usage{
NTE.plotDeposition(d, buffer = 0.5)
}
\arguments{
\item{d}{data frame with NTE data}

\item{buffer}{how much data to show before / after plot, 1=deposition time}
}
\value{
ggplot graph
}
\description{
graph of the deposition, colors the actual deposition portion
}
\examples{
d = NTE.load('20191121.txt')
g1 = NTE.plotDeposition(d)
}
