% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NTE.plotSummary.R
\name{NTE.plotSummary}
\alias{NTE.plotSummary}
\title{makes 4 plots with NTE data}
\usage{
NTE.plotSummary(d, sz = 2)
}
\arguments{
\item{d}{data frame with NTE data}

\item{sz}{size of data points in graphs}
}
\value{
ggplot graph
}
\description{
makes 4 plots with NTE data
}
\examples{
d = NTE.load('20191121.txt')
}
